pyproject.toml
[tool.poetry]
name = "poetry shell"
version = "0.1.0"
description = ""
authors = ["Dmitriy Kucheriaviy <dmitask@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
django_settings_module = "config.settings"